feat: enhance coach and insights routers, improve context creation, a…

+++ apps/server/ai-service/src/routers/insights.router.ts
@@ -11,7 +11,9 @@ export const insightsRouter = router({
   health: publicProcedure.query(() => ({ ok: true, service: "ai-service" })),
   list: protectedProcedure
     .input(
-      z.object({ limit: z.number().min(1).max(50).default(10) }).optional(),
+      z
+        .object({ limit: z.number().int().min(1).max(50).default(10) })
+        .optional(),
     )
     .query(async ({ ctx, input }) => {
       const result = await AIService.getInsights(
